* FIX: xrLC debug configuration building * FIX: xr_aligned_malloc from warning * FIX: Crash, because xrLC_GlobalData::create_rcmodel invoked with already converted CDB:TRI

+++ code/engine.vc2008/utils/xrLC/xrBuildRapidModel.cpp
@@ -93,7 +93,7 @@ void CBuild::BuildRapid		(BOOL bSaveForOtherCompilers)
 		{
 			F->flags.bProcessed	= true;
 
-			CL.add_face_D		( F->v[0]->P,F->v[1]->P,F->v[2]->P, *((size_t*)&F), F->sm_group );
+			CL.add_face_D		( F->v[0]->P,F->v[1]->P,F->v[2]->P, *((u64*)&F), F->sm_group );
 		}
 	}
